+++ src/rules/IfStatementBraces.js
@@ -29,12 +29,17 @@ const rule = {
                 return allLines;
             }
             let line = allLines[lineNumber];
+            // Check we are on the correct line to correct, if not trigger error
             if (!line.includes("if")) {
                 throw new Error('Line does not contain "if" :' + line);
             }
+            if (line.includes("{")) {
+                throw new Error("Line already has an opening brace :" + line);
+            }
+            // Add opening brace to if line
             line = line.trimEnd() + " {";
             allLines[lineNumber] = line;
-            // next line
+            // Add a tag ###NEWLINECLOSINGBRACE### to indicate rule ClosingBraceNotAlone to replace it by a closing brace & carriage return
             let match = false;
             let pos = 0;
             let level = 0;
@@ -81,6 +86,60 @@ if (new File(sfdxWorkingDir + '/.sfdx').exists() && this.promptForReloadMetadata
 }
 else
     doRetrieve = true
+`
+        },
+        {
+            sourceBefore: `
+if (allowCreation==true) {
+    String scratchOrgSlctn ;
+    if (this.scratchOrgUserEmail == null) 
+        scratchOrgSlctn = Utils.userInputSelect('User input','Please select a scratch org number , or 0 to create a new scratch org : ',orgsChoiceList, 5);
+    if (scratchOrgSlctn != null && scratchOrgSlctn != '' &&
+        orgsChoiceMap[scratchOrgSlctn] != null && orgsChoiceMap[scratchOrgSlctn].alias != null &&
+        orgsChoiceMap[scratchOrgSlctn].alias != '' && this.scratchOrgUserEmail == null) {
+            // Select scratch org
+            this.scratchOrgAlias = orgsChoiceMap[scratchOrgSlctn].alias ;
+    } else {
+            // Create new scratch org
+            if (this.scratchOrgUserEmail == null) {
+                this.scratchOrgAlias = Utils.userInputText('Please enter the name of the new scratch org (without spaces or special characters, AND WITH YOUR NAME IN IT FOR GOD SAKE :) ', 5) 
+                // Store choice if request in config file 
+                if (Utils.userPromptOkCancel('Do you want this new scratch org to be your default one ? (in '+this.ownConfigFile+')', 5)) {
+                    Utils.setPropInJsonFile(this.ownConfigFile,"scratchOrgAlias",this.scratchOrgAlias)
+                }
+            }
+            // Define scratch org description as JSON
+            if (true)
+                this.defJsonCreation()
+    }
+}
+`,
+            sourceAfter: `
+if (allowCreation==true) {
+    String scratchOrgSlctn ;
+    if (this.scratchOrgUserEmail == null) {
+        scratchOrgSlctn = Utils.userInputSelect('User input','Please select a scratch org number , or 0 to create a new scratch org : ',orgsChoiceList, 5);
+    }
+    if (scratchOrgSlctn != null && scratchOrgSlctn != '' &&
+        orgsChoiceMap[scratchOrgSlctn] != null && orgsChoiceMap[scratchOrgSlctn].alias != null &&
+        orgsChoiceMap[scratchOrgSlctn].alias != '' && this.scratchOrgUserEmail == null) {
+            // Select scratch org
+            this.scratchOrgAlias = orgsChoiceMap[scratchOrgSlctn].alias ;
+    } else {
+            // Create new scratch org
+            if (this.scratchOrgUserEmail == null) {
+                this.scratchOrgAlias = Utils.userInputText('Please enter the name of the new scratch org (without spaces or special characters, AND WITH YOUR NAME IN IT FOR GOD SAKE :) ', 5) 
+                // Store choice if request in config file 
+                if (Utils.userPromptOkCancel('Do you want this new scratch org to be your default one ? (in '+this.ownConfigFile+')', 5)) {
+                    Utils.setPropInJsonFile(this.ownConfigFile,"scratchOrgAlias",this.scratchOrgAlias)
+                }
+            }
+            // Define scratch org description as JSON
+            if (true) {
+                this.defJsonCreation()
+            }
+    }
+}
 `
         }
     ]